feat: Style download button and move it to its own section. #7
This commit improves the UI for downloading the avatar as an PNG image.

The Download button is now placed in a dedicated section below the footer to fit nicely in the current structure of the UI.
Furthermore the button is styled in such a way that it fits to the rest of the UI and is visible as a clickable button on hover.